Fix - Increased jwt token expiration timeout to 4 hours

+++ config/jwt.php
@@ -101,8 +101,7 @@ return [
     |
     */
 
-    'ttl' => env('JWT_TTL', 60),
-
+    'ttl' => (int) env('JWT_TTL', 60),
     /*
     |--------------------------------------------------------------------------
     | Refresh time to live
@@ -120,8 +119,7 @@ return [
     |
     */
 
-    'refresh_ttl' => env('JWT_REFRESH_TTL', 20160),
-
+    'refresh_ttl' => (int) env('JWT_REFRESH_TTL', 20160),
     /*
     |--------------------------------------------------------------------------
     | JWT hashing algorithm
@@ -132,7 +130,6 @@ return [
     */
 
     'algo' => env('JWT_ALGO', 'HS256'),
-
     /*
     |--------------------------------------------------------------------------
     | Required Claims
@@ -205,8 +202,7 @@ return [
     |
     */
 
-    'leeway' => env('JWT_LEEWAY', 0),
-
+    'leeway' => (int) env('JWT_LEEWAY', 0),
     /*
     |--------------------------------------------------------------------------
     | Blacklist Enabled
@@ -232,8 +228,7 @@ return [
     |
     */
 
-    'blacklist_grace_period' => env('JWT_BLACKLIST_GRACE_PERIOD', 0),
-
+    'blacklist_grace_period' => (int) env('JWT_BLACKLIST_GRACE_PERIOD', 0),
     /*
     |--------------------------------------------------------------------------
     | Cookies encryption
